Update namespaces after moving files
