+++ src/Infrastructure/Publishers/DevToPublisher.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace N3XT0R\XPub\Publishers;
+namespace N3XT0R\XPub\Infrastructure\Publishers;
 
-use N3XT0R\XPub\Contracts\PublisherInterface;
+use N3XT0R\XPub\Domain\Contracts\PublisherInterface;
 
 class DevToPublisher implements PublisherInterface {
     public function publish(string $title, string $content): bool {
