+++ src/Infrastructure/Wordpress/Logging/Handler/AdminNoticeHandler.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace N3XT0R\XPub\Infrastructure\Logging\Handler;
+namespace N3XT0R\XPub\Infrastructure\Wordpress\Logging\Handler;
 
 use Monolog\Handler\AbstractProcessingHandler;
 use Monolog\Level;
