+++ src/Infrastructure/Wordpress/Logging/Handler/WPDBLogHandler.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace N3XT0R\XPub\Infrastructure\Logging\Handler;
+namespace N3XT0R\XPub\Infrastructure\Wordpress\Logging\Handler;
 
 
 use Monolog\Handler\AbstractProcessingHandler;
